Better organize code
