+++ Sources/Player/FailingResourceLoader.swift
@@ -1,0 +1,30 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import AVFoundation
+
+final class FailingResourceLoaderDelegate: NSObject, AVAssetResourceLoaderDelegate {
+    private let error: Error
+
+    init(error: Error) {
+        self.error = error
+    }
+
+    func resourceLoader(
+        _ resourceLoader: AVAssetResourceLoader,
+        shouldWaitForLoadingOfRequestedResource loadingRequest: AVAssetResourceLoadingRequest
+    ) -> Bool {
+        loadingRequest.finishLoading(with: error)
+        return true
+    }
+
+    func resourceLoader(
+        _ resourceLoader: AVAssetResourceLoader, shouldWaitForRenewalOfRequestedResource
+        renewalRequest: AVAssetResourceRenewalRequest
+    ) -> Bool {
+        true
+    }
+}
