+++ Sources/Player/LoadingResourceLoader.swift
@@ -1,0 +1,23 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import AVFoundation
+
+final class LoadingResourceLoaderDelegate: NSObject, AVAssetResourceLoaderDelegate {
+    func resourceLoader(
+        _ resourceLoader: AVAssetResourceLoader,
+        shouldWaitForLoadingOfRequestedResource loadingRequest: AVAssetResourceLoadingRequest
+    ) -> Bool {
+        true
+    }
+
+    func resourceLoader(
+        _ resourceLoader: AVAssetResourceLoader, shouldWaitForRenewalOfRequestedResource
+        renewalRequest: AVAssetResourceRenewalRequest
+    ) -> Bool {
+        true
+    }
+}
